requirements.txt
pandas
biopython
numpy
scikit-learn
scipy